pyproject.toml
[project]
name = "support-triage-agent"
version = "0.1.0"
description = "AI Support Ticket Triage Agent using FastAPI, LangChain, and Gemini"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.29.0",
    "python-dotenv>=1.0.0",
    "langchain>=0.1.0",
    "langchain-google-genai>=1.0.0",
    "langchain-chroma>=0.1.0",
    "langchain-community>=0.0.20",
    "pypdf>=4.0.0",
    "reportlab>=4.0.0",
    "pydantic>=2.6.0"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
